requirements.txt
# Core Django
Django==4.2.21
djangorestframework==3.16.0

# Database
psycopg2-binary==2.9.10

# Media and File Handling
Pillow==11.2.1
pilkit==3.0
django-imagekit==5.0.0

# Content Management
django-ckeditor==6.7.2
django-js-asset==2.2.0
django-taggit==6.1.0

# SEO and Meta
django-meta==2.5.0
django-seo==0.3.5
django-sitemaps==2.0.1
django-robots==6.1

# Performance and Compression
django-compressor==4.5.1
django-appconf==1.1.0
rcssmin==1.1.2
rjsmin==1.2.2

# Development Tools
django-extensions==4.1
django-debug-toolbar==5.2.0

# Background Tasks
celery==5.5.2
redis==6.2.0
kombu==5.5.3
billiard==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
async-timeout==5.0.1
vine==5.1.0

# Utilities
python-decouple==3.8
python-slugify==8.0.4
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0

# Production
whitenoise==6.9.0
lxml==5.4.0

# Text Processing
text-unidecode==1.3
wcwidth==0.2.13

# Type Checking
typing_extensions==4.13.2

# Time Zone
tzdata==2025.2

# Template Processing
amqp==5.3.1
asgiref==3.8.1
prompt_toolkit==3.0.51
sqlparse==0.5.3